theme switch moved to settings page
